requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
requests
google-generativeai
toml
tornado==6.3.3